# code/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "I have no specific description",
        "title": "Code API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/v0",
    "paths": {
        "/charge": {
            "post": {
                "description": "create charge",
                "tags": [
                    "Charge"
                ],
                "summary": "create charge",
                "operationId": "create_charge",
                "parameters": [
                    {
                        "description": "charge code info",
                        "name": "charge_code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputs.ChargeCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/charge/{charge_code}": {
            "get": {
                "description": "get charge",
                "tags": [
                    "Charge"
                ],
                "summary": "get charge",
                "operationId": "get_charge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "charge code name",
                        "name": "charge_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/charge/{charge_code}/apply": {
            "post": {
                "description": "charge code",
                "tags": [
                    "Charge"
                ],
                "summary": "charge",
                "operationId": "charge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "charge code name",
                        "name": "charge_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/charge/{charge_code}/rollback": {
            "post": {
                "description": "charge rollback",
                "tags": [
                    "Charge"
                ],
                "summary": "charge",
                "operationId": "charge_rollback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "charge code name",
                        "name": "charge_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "inputs.ChargeCode": {
            "type": "object",
            "properties": {
                "expiration_date": {
                    "type": "string"
                },
                "max_capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}